Modify _form_basics.html.erb so it's balanced
Rails ERB templates do not need to have balanced HTML (e.g., you can start
a <div> without having a matching </div> in the same template).
However, it's easy to have bugs if an HTML structure
is started in one template, but completed in another, and they're
harder to debgu.  So where practical,
it's better to make templates balanced (even if a template
has a few more lines than needed).  This commit removes a few lines
from _form_basics so that it's balanced (and the callers will be too).