Modify level 1&2 so N/A allowed & justifications required in many criteria
For level 1 (silver) and level 2 (gold) it makes sense to be far
more demanding and require justifications for basically everything.
In addition, in many places N/A would be valid answer, and we didn't
always include it.  Modify the criteria.yml file to handle this.
Note: We can remove justification text requirements in some fields later
if we want to; it's easier to remove those requirements later
than to add justification requirements later.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>