Add tests: exactly correct criteria displayed, and ensure Criteria has keys
This commit improves the tests.

For each level of the badge entry show view, this commit checks to ensure
that every criterion that is *supposed* to be in this level is selectable,
and that every criterion that is *not* supposed to be in this level
is not selectable.  That way, if a level incorrectly includes or omits
something, it will be detected during testing.  We don't need to separately
test the badge entry edit view, because they call the same code.

We also add a test to ensure Criteria is loaded with a
reasonable number of criteria (keys) at each level.
This is an overall sanity check for size.  Many other tests loop through
criteria, but those tests will quietly succeed if the loop has 0 length.
Here, we make sure we *have* lengths, so that the other tests are actually
testing something.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>