Add UI for changing additional rights. Fixes #91
Add a simple interface for editing additional rights.
This makes it possible to create groups for editing
without GitHub (e.g., for PostgreSQL).

Note that only users who control a project (the owner
or a BadgeApp admin) can change the list of additional users
who can edit a project entry.

The interface requires that changes begin with "=";
this simple confirmation trick should reduce the risk
of accidentally changing it.  Users can set it to an
empty list with a simple "=".  The "details" text
provides more details for those who care.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>