Put radio buttons in a fieldset with legend for accessibility
Radio buttons are supposed to be in a fieldset with legend.
This is for WCAG 2.0 (Level A), Guideline Group 1.3,
"Ensure that information and structure can be separated from presentation"
and noted in:
https://achecker.ca/checker/suggestion.php?id=168

I had to forcibly make legend "display:none !important".
This isn't ideal, but everything else I tried produced terrible results
for typical users.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>