Remove stress markings in the name of the Russian language
Remove stress markings in the Russian name for the Russian language.
My thanks to Yannick Moy for the report!

According to the Institute of Russian Language of the Russian Academy
of Sciences, an optional acute accent may, and sometimes should,
be used to mark stress (e.g., to eliminate ambiguity).
(Source: https://en.wikipedia.org/wiki/Russian_language#Orthography ).
However, in this case there's no ambiguity, and the text isn't
here for beginners. In this situation the stress markings are
normally not included.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>