Modify many database fields to be "not null"
This prevents some data problems, since even if the application
has a bug, the database will prevent certain invalid values.

For unknown status values we use '?' and set it as the default,
instead of using "null" to indicate unknown values.
That way, the value displayed to the user in English
is exactly the same as what we store.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>