Set up Criteria model to look for criterion details in lower levels
If a criterion exists on multiple levels, the details will
in general not change.  This allows us to look for details
for a higher level criterion in the lower levels, reducing
duplication in the various translation files.

We have removed all duplicate details in en.yml except if the lower
level criterion is a future criterion.  This will allow us to
quickly remove the future criteria in the future without losing
the details text on higher levels.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>