Ensure that view caches depend on locale when relevant
Most view caches now need to depend on locale as at least one
of their inputs, since the text displayed depends on that.
However, in our reorganization we didn't always include them - fix that.
An exception is JSON data (the JSON keys are fixed values regardless
of the locale, so the cache does *not* need to depend on the locale
for JSON data).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>