Add explanation of project_stats new and edited projects; fixes #778
GeorgLink recommended that we explain these datasets better
(it's hard to translate text unless it's clear what it means).
This adds some text to explain it further.

He also noted that it can be hard to see the details due to overlap; this
is true, but the charts' ability to hide/unhide data makes it convenient.
However, it's only convenient if you know it can do that,
so this text also mentions that you *can* do that.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>