Raise error during test if no corresponding i18n key is available
Raise an error during testing if we are asked for an i18n key,
but there is no corresponding i18n key is available.
There is a backoff to English if no translation in the current locale
is available, so this error is only raised if the key is not
in the English text.

This is still a significant requirement; it means
that all keys must be found in all pages retrieved during the tests,
and since we work for good coverage, it means that the i18n keys
are given a workout.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>