Remove rails_db from development environment
Remove rails_db from the development environment, as it's no longer as useful.

The rails_db gem doesn't like our CSP environment.
We could work around that, but in practice now we typically use other
tools to access the database (such as interactive "rails db"), so
there is less call for the rails_db capabilities.
Also, by removing this, the end of the file config/routes.db is also
the end of all defined routes, which makes some route definitions easier.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>