In ProjectsController#badge select only the fields needed (less memory)
Select only the fields we need on a badge request,
so that its ActiveRecord will require less memory.
Scout reports that ProjectsController#badge is a significant
requester of memory, and this reduces its requests.
Strictly speaking we don't need "id", but it's useful to have it for
debugging & including it costs very little.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>