Add tests to check that common pages accept 'get' in every locale
Get common pages in every locale to see if the locale works.

Some locales have additional needs, e.g., Russian requires more keys
than English to handle pluralization. Failing to provide the keys
will cause the entire page to fail to load.

This does *not* require that the text be translated to all languages.
If the text is not translated, English is used.
But this test *does* require that all required keys be present
(in the locale or in en.yml) so that we *can* do the translation later.

This is a regression test.  When we added Russian, we didn't add the
necessary keys to handle Russian pluralization, which meant that /projects
would not load.  From now on we will automatically *notice* if a common
page fails in any permitted locale.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>