Add class to criteria that can be hidden when viewing show.
Instead of evaluating which criteria can be hidden on show (not edit)
lets just precompute these criteria and add a class flagging as such.
This will make ensure that only criteria which are qualified to be
hidden will be (those that have the result :criterion_passing on
the ruby side).  There was a possibility that on show this fuctionality
would not work properly due to how we handle markdown (everything worked
fine when editing).  This commit ensures we are always consistent.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>