Make locale selection a dropdown in the navigation bar (#809)
* Make locale selection a dropdown in the navigation bar

The selector reloads the current page in the new locale
(including the same query).  For some reason this doesn't
include the URL anchor (that's less important, to be fair).

This uses Rails caching to speed page generation and reduce
the number of objects created.  The selector is loaded a
part of *every* page, including common pages, and the selector
has to loop through a number of options.

* Bugfix: Generate non-en locale in root page in another non-en locale

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>