Add stub translations for Japanese & Russian on home page & layouts
These are translations from Google Translate, with a few light fixes
to repair the HTML and make the links point to the correct locale.

These are not good human translations. Instead, they are important
stubs we need for testing.  These translations
make it easy to determine that when we request Russian or
Japanese we're actually *getting* that locale.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>