Speed up some of the feature tests.
The feature tests tended to be very slow. One of the main
reasons for this was the use of assert has_content?('id')
when looking to see if a particular criterion was visible.
This is very slow as since no anchors were specified the
entire document had to be parsed.  Instead we now use
assert/refute_selector(:css, '#id'), which is much faster
since an anchor is specified.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>