Add (translated) pop-overs for silver & gold to clarify purpose
In other locales we want the *images* to stay the same, but to also
make it easy for people to understand what they mean.
Simple solution: translated pop-overs.  Just hover over the image,
and the translation will appear.  The images already have ALT text, but
that's for accessibility - normally-sighted users might not notice that.
This solution also means we don't have to ask the translators to
translate any more keys.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>