Enforce in DBMS that local emails are unique. Fixes #719
Modify the schema so that the *database* enforces uniqueness
of email addresses when the provider is 'local'.
We depend on this, because we use the email address as the
identifier during login.

We enforce this at the Rails level, but that is subject to a race
condition because different Rails processes or threads might not
see this.  Also, enforcing important data structure requirements
at the DBMS level means that even if the code has a bug, it can't
violate this condition.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>