Allow download of user index by non-admins
We take privacy seriously, so in the past we have only allowed admins
to get the user index.  However, in retrospect this was really
more security theater than real security.  The index information
is available publicly anyway through various means.
The user information (id and name) is easily determined via show
one-by-one.  Whether or not a user is on GitHub, and their nickname,
can be inferred by the external page reference.  And the GitHub uid
is easily determined from GitHub using the nickname and GitHub APIs.
So in short, we're overprotecting information already publicly available.

This is a problem, because we *want* to encourage people to analyze
our data about OSS, and the user list is a key part of that data.
We're just making it hard for legitimate uses.

So, this lets download the user index even if they aren't an admin.
It's still subject to the usual limitations (e.g., pagination).

This change still requires that users be logged in to get the data.
This seems fair enough (register yourself before getting a list
of users), and may help us detect abuses.

The tests use "refute ..." instead of "assert !..." per a suggestion
from Jason Dossett.

Note that we still *ONLY* release email addresses to admins, and
we NEVER include the hashed passwords. Those remain sensitive,
and we continue to protect them per our security policy.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>