Document how to download data for analyis.  Fixes #822
Other researchers will want to download the project data
for further analysis.  Yannick Moy, in particular,
wanted to analyze the data further.  We want to encourage
research - the more analysis that can occur, the more everyone
can understand FLOSS better.

We already support download, but it wasn't as clearly documented.

This commit documents exactly how to download the key project data.
Of course, we do *not* provide some data
(like user email addresses and hashed user passwords),
and this is intentional - so we document that too.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>