Record statistics for higher-level badges.
This partly addresses #813 by recording data in table project_stats.
This commit does not *display* any of this data - but we
have to record it before we display it.

It's not clear we'll report the 25%, 50%, 75%, or 90% levels for
silver or gold, but it's easier to simply start recording it & then
we can show subsets if desired.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>