Include additional rights in projects JSON file
We want people to be able to analyze our data.  The "additional rights"
table is a separate thing - include it in the projects' JSON file,
so that it can be extracted.  In most cases it'll be empty.
Generate an array of hashes - that's more flexible, because new fields
can be easily added to the hash.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>