Cache justification markdown
Cache the generated markdown for project justifications.
Note that these can be *reused* across locales, and when they are
identical these can even be reused across projects
(this can happen when people accept the system-created justifications).
The markdown is generated as part of file
app/views/projects/_status_chooser.html.erb, which is the biggest
user of time and memory allocations, so optimizing this file
is sensible in general.

We use the justification text as the cache key.
This can result in rediculously-long cache keys, but those are exactly
the cases where the cache would be most useful, so we permit it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>