Add test to ensure semvar description is not escaped
Jason Dossett noted in his review of commit d3906aee63c4b
that we didn't have a test to ensure that description text
is not escaped (it is not supposed to be escaped, but no
test checked).

This commit adds a test to ensure that a description with HTML is
not HTML escaped in the response body.  We don't have that problem
now, and we'd like to keep it that way.  This way, the description
text can include markings like emphasis and hypertext links.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>