Split up caches in status_chooser
The status_chooser is a performance hotpoint in this applciation.
Split the caches into multiple caches, so its results are more
likely to be shared, and completely eliminate the use of
the justification text as a cache key
(this text can be long, make it less than ideal as a cache key).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>