Add summary of silver & gold to README. Fixes #842 (#846)
* Add summary of silver & gold to README. Fixes #842

We already had a summary of the "passing" criteria in the README.
Here we add summaries for silver & gold.  These summaries aren't the
complete list; they simply give users a quick idea of what the
criteria are like.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* * Remove double-quotes in link text - markdownlint workaround

Markdownlint has rule MD039 "spaces inside link text"; unfortunately,
this rule incorrectly triggers when there are double quotes in link text.
Last I checked, double quotes are not spaces.  However, it's easy
to work around this, so just work around it instead of trying to
fix the markdownlint tool.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>