Test: All text values (all locales) incl. only acceptable HTML
This commit adds a test to ensure that all text values,
in all locales, include only acceptable HTML.  We determine this by
sanitizing (which only accepts a whitelist of tags and attributes),
and then seeing if that's the same as simply regularizing the text
(without removing anything).

If we pass the test, then we can omit sanitization or escaping of
the locale text at run-time with confidence, since we know
that it can only include the permitted tags and attributes.
We also know that *all* the text has been checked ahead-of-time.
That's better for performance (at run-time we definitely don't
need any of those checks) and for error detection (we detect
an unexpected tag or attribute much earlier, and more thoroughly).
Strictly speaking the locale text is trusted anyway, but
adding a test to check for problems seems reasonable.

This commit does add a non-trivial amount to test time.
This one test takes about 10 seconds locally.
But while that longer test time is not ideal, it's not
a show-stopping amount of test time.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>