Test: All text values (all locales) incl. only acceptable HTML (#850)
* Test: All text values (all locales) incl. only acceptable HTML

This commit adds a test to ensure that all text values,
in all locales, include only acceptable HTML.  We determine this by
sanitizing (which only accepts a whitelist of tags and attributes),
and then seeing if that's the same as simply regularizing the text
(without removing anything).

If we pass the test, then we can omit sanitization or escaping of
the locale text at run-time with confidence, since we know
that it can only include the permitted tags and attributes.
We also know that *all* the text has been checked ahead-of-time.
That's better for performance (at run-time we definitely don't
need any of those checks) and for error detection (we detect
an unexpected tag or attribute much earlier, and more thoroughly).
Strictly speaking the locale text is trusted anyway, but
adding a test to check for problems seems reasonable.

* Improve test error message

If the test fails, make it easier to find where things went wrong.

* Massively speed up test for locale text

Speed up test for locale text.  It turns out that most strings
don't have "<" in them at all, and by definition such strings
cannot have tags or attributes. So when we see such strings,
skip the more detailed test if by definition the test will succeed.

* Rename variables to make their purpose clearer

Per review feedback from Dan Kohn.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>