Recursively check locale text for common mistakes
Recursively check locale text, e.g., ensure that it has acceptable HTML.
Among other things, this is a regression test to prevent the problem
fixed in commit 48aa4999d3d81c4, where a Cyrillic "a" was used instead
of a Latin "a" in a closing "/a".

We pass parameter "from" so that if there's a problem,
we can report exactly where the problem comes from
(making it easier to fix).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>