Modify locale-relate Rake tasks - check, autofix if needed
When translation:sync is run, check the results.
Then, if there's a problem, try to fix it.

Doing things this way means that we do *not* hide problems
(we made them visible), while simultaneously trying to fix
them for users (so that we don't have to rush to fix
common problems).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>