Add prerequisite criteria for higher levels; Fixes #838
This commit officially adds the achieve_passing and achieve_silver
criteria for the silver and gold levels respectively.
The statuses for these criteria automatically calculated and stored
in the database along with an empty justification. This implementation
was chosen as it makes the code a bit simpler.

Upon deployment, the rake task
"update_all_higher_level_badge_percentages" must be ran to update
the new fields for all projects.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>