Fix caching in status_chooser; fixes #885 (#886)
The "buttons" cache in _status_chooser was no longer  storing
any information on the criterion_result (which is influenced by
the justification text).  This was causing the result_symbol
for various projects to display incorrectly.

This is hard to set up a regression test for, so I have not
made one.  We could perhaps make a long feature test where
we change a criterion's justification text, save and compare
with expected results.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>