Add user interface for editing project additional rights. Fixes #93
This adds a user interface so that users can allow *additional*
other users to edit project entries, beyond those already allowed
(owner, BadgeApp admins, and those who can edit that GitHub repo where
that applies).

We expect that in nearly all projects this functionality
will not be used, but that when it's needed it's important.
We've contemplated this functionality for a while,
but most projects haven't asked for it.  However, the PostgreSQL project
informed us that this functionality was important to them, and that
created a clear situation where there *was* a need.
Since this is a rare but important need, this commit implements
a minimal UI that "gets the job done".

The authorization rules implemented here are straightforward:
* If you control a project (are the owner or a BadgeApp admin), you can
  remove users from the additional_rights list.
* If you can edit a project, you can add other users to the project
  with additional rights.  You can edit a project if you control that project
  *or* are on that project's additional_rights list of users.
  This means that people can add their friends without trouble, but
  the owner of the badge entry can remove everyone else if there's a problem.

This version has separate "add" and "remove" subcommands, instead of
simply setting the value to a list (as an earlier version did).
There are two reasons for this:
1. It avoids race conditions.
2. Removing users seems more privileged than adding them, so this
   approach makes it easy to be more stringent on who can remove users.

We could have only allowed adding or removing a single user at a time,
but it's not much harder to support a list.  Since lists can be useful,
we support lists.

The tests specifically include a security check: a users on the
additional_rights list cannot remove people from the list.
In the future we may allow people to remove themselves, but for now
that seems unnecessary; the owner or a BadgeApp admin can remove people.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>