Partial implementation: Automatically determine locale
We want users to automatically see their preferred locale.
This changes how URLs are interpreted; "no locale in URL"
is now interpreted as "redirect based on browser preference"
and not as "use English".

This is only a partial implementation.  A large number
of tests fail, because most tests don't specify a locale
(and thus get redirected).  Some tests fail for other reasons.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>