Set updated_at if additional_rights changes. Fixes #912
Ensure that a project entry's updated_at time is set even when
a project only adds/removes editors (in the additional_rights)
without changing anything else in the badge entry.
In a sense this "invalidates" the cache, though we actually don't
invalidate caches, we just change the values of the key used to access
the cache.  One of the values used in the key is the project's
updated_at value.

Previously the update time was unchanged in this case, because
we updated the editor list *after* we saved the project entry.
As a result, users saw obsolete cached values.

Instead, this commit reorders the actions so that we first
update the editor list, and *then* save the project entry.
Rails will automatically update the project entry when the editor
list is changed, but the previous code was ordered in such a way
as to discard this data.  In this code reordering the new (correct)
updated_at value is saved.

This commit only moves one line in the runtime.  The rest of the changes
are a regression test to prevent recurrence of the bug.
We do by adding a test to ensure that the update time is updated in this case.
This kind of bug can only happen when data is updated in *other*
tables, and that is unusual in this particular application.

This bug only impacts projects that edit the list of editors (we
only have 1 in almost 1000), and that this only temporarily makes
the display incorrect (any other edit will fix it, as will the daily
cache resets). The actual actions (e.g., who is allowed to edit) are
not affected - this only affects what is displayed. So this is not a
serious bug. Still, it is a bug, so this commit fixes it.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>