Add pickier checks on SPDX license expression on client, fixes #916
Check that only valid characters are included in
a SPDX license expression on the client.

This isn't a rigorous check, but it eliminates some nonsense.
This commit only checks on the client side (using HTML5 patterns);
we don't need to do more, because invalid SPDX license expressions
aren't a security vulnerability.  Instead, we simply want to
make it a little harder to enter data that is clearly nonsense.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>