Add more licenses to SPDX license dropdown, based on projects pursuing badges
An analysis of the licenses of the projects currently pursuing badges
showed that a few licenses used in those projects aren't in the dropdown.
This isn't a crisis, since projects can manually fill in this information
(and in some cases the autofill can figure it out).
However, if people fill in the SPDX license expression by hand
there's an increased likelihood of error.
That's especially true for this field, because many developers aren't
familiar with SPDX license expression syntax.

This commit adds dropdowns for a few more licenses that are being
used by projects currently pursuing badges.  It also fixes
a spelling error in an existing license.  This should increase
the likelihood of getting correct data.

The capitalization of the zlib-related licenses is odd, but it's straight
from the SPDX license list at <https://spdx.org/licenses/>.
Since there's a spec, it seems reasonable to follow it.
SPDX licenses aren't case-sensitive anyway, so the result is correct
no matter what its case is.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>