Update transitive dependency gems
This updates:

- autoprefixer-rails (7.1.2.6->7.1.3)
- coderay (1.1.1->1.1.2)
- sprockets-rails (3.2.0->3.2.1)

To make this work, we also have to expressly tell rails_best_practices
that 'Criteria#met_justification_or_url_required?' is used.
Unfortunately, the rails_best_practices tool can't determine
when a method is used in an ERB template, so methods that are only used
in ERB templates (like this case) trigger false positives.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>