Add support for justification text comments. Fixes #929 (#930)
* Add support for justification text comments. Fixes #929

Up until now justification text is used in two different ways
when a criterion isn't met:

1. Text that justifies why it's acceptable that something isn't met.
2. Commentary that explains why someone (typically an automated
   Detective/bot) has determined that it's not met, but with
   no argument that this is acceptable.

However, we currently can't distinguish between those two cases. As a
result, when a detective adds justification text explaining "why we've
determined this is not met" (way 2), the rest of the system interprets
it as a justification of why it's okay (way 1).

We need to distinguish these two different approaches, hopefully in a
way that's easy to implement & use.

This commit changes things slightly.
If the first character in a justification is "#", we interpret the text
as "commentary" (way 2), and do NOT count the text as a justification for
why this is acceptable. The "#" character is widely used as a comment
symbol in many programming languages. This approach is easy to implement,
and should be fairly easy to use as well (in particular, I think it'd
be easier than adding a badge of new UI widgets).

Implementing the change in badge calculation is trivial; almost all
the changes are to the detectives so that they can report this new case.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* In justification comments, ignore URLs and add tests

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Minor code reformat to make Rubocop happy

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* In JavaScript handle "undefined" specially for justification text

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Use regex instead of startsWith to resolve circleCI test issues

Tests on circleCI were failing with the use of startsWith in js.
We are not sure why but switching to regex resolves the test failures
so we switch to that instead.

Signed-off-by: Jason Dossett <jdossett@utdallas.edu>