Update gem rubocop (0.49.1->0.51.0); replace .dup with (+...) to unfreeze
Update the Rubocop gem.

This new version of Rubocop complains if you use .dup
to unfreeze a string, and wants you to use prefix "+" to unfreeze a string.
However, if you just use prefix "+", that triggers Lint/AmbiguousOperator.
The recommended solution is to use the form METHOD_NAME(+'frozen_string',...);
that is oddly inconsistent, but it works, so we'll use it.

All these uses are for the first parameter of assert_select when
it has more than one parameter.  This use really should require
unfrozen strings anyway, as discussed here:
https://github.com/bbatsov/rubocop/issues/4630
So in the future we may not need to do this workaround, but we must
do something for now.  This workaround works *now*, and it
won't be *wrong* in the future, so that's fine.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>