Hardening: Move CSRF meta tags to very early in application HTML
Move the CSRF meta tags so they are very early in sent HTML.
This is a hardening measure; this way, even if there's an
error in the later code (e.g., that allows an attacker to insert a
comment or other text), it cannot override or control the CSRF meta tag
because the CSRF meta tag occurs first.

For more about this, see the "Dangling markup injection" section of
"Postcards from the post-XSS world" by Michal Zalewski:
http://lcamtuf.coredump.cx/postxss/

This does have a trivial downside: it interferes slightly with compression,
because now we have random data inserted somewhat early.
This isn't serious, indeed, it's unlikely to be noticed by actual users,
and we highly rank security.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>