Note in security.md that we apply negative testing (test what shouldn't do)
A common mistake in testing is neglecting to test what the software
should NOT do.  We *do* test that many security-related things that
should *not* happen actually don't happen, so note that.

For example, we test that the software does NOT let in users with the
wrong password, that normal users CANNOT edit others' projects, and that
normal users CANNOT see email addresses on user pages.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>