Show additional rights and admin privilege in user page. Fixes #915
On the user page show the user's "additional rights".
These determine the specific rights of
a user, but up to now have not been shown on the user page.
Show them, so that it's completely obvious what rights the user does
(and does not) have.

If you're a logged-in admin, also show on the user's page
if that user is an admin.  It's probably not hard to figure out
who at least some of the administrators are, but this is data
about *people* instead of system *design*.

Add tests, including a number of negative tests
to verify that this data is *only* shown to admins, and document
these negative tests in security.md (we expressly note that we do
negative tests, and this is another good example of this).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>