Require URLs for a few more "gold" criteria. Fixes #926.
The "gold" criteria require justification for practically everything,
instead of requiring URLs for justification.  Since "gold" is a high
standard, it's reasonable to ask for more information (URLs) in
the justification.  One of the main reasons to ask for URLs is to
make it easier for others to verify claims made; it also makes it
easier to look at those badge holders as examples of "how to do it".

Ignoring the criterion "must meet previous level", this means that when
you examine the percentage of criteria that require URLs in justifications,
gold requires a higher percentage (41%=9/(23-1)) compared to silver
(31%=17/(55-1)).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>