Add info about cookies to comply with EU and UK "cookie laws". Fixes #938 (#952)
The EU and UK have "cookie laws".
More information about the EU cookie laws, from a formal EU site, are here:

* [(EU) Information Providers Guide - Cookies](http://ec.europa.eu/ipg/basics/legal/cookies/index_en.htm)
* [ePrivacy directive, especially Article 5(3)](http://eur-lex.europa.eu/LexUriServ/LexUriServ.do?uri=CELEX:32002L0058:EN:HTML)

It's not clear that these laws applies to the site:  I'm a US citizen,
this site is run in the US, and the primary organization running the
site (the Linux Foundation) is a US organization.  In addition,
[NoCookieLaw](http://nocookielaw.com/) has a valid point: these laws seem
to be completely pointless, since they doesn't actually provide any *real*
privacy (and ignore the *real* and serious privacy problems).

However, in our case, it looks like all we *need* to do is to add a
little text and connect it what was named "Remember Me" to
explain the basics of what it does.  That doesn't appear to be hard to do.

So instead of wondering if we really need to comply or not, we'll just
make an honest effort to comply, which is what this commit attempts to do.

The text is based on "cookie-notice-template.zip"
file "cookie-notice-template_en.doc", which additions that attempt
to explain the specific circumstance here.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>