Rescale project logo image widths on home page
Provide precomputed width values for the project logo images
on the home page.  By precomputing, we remove image loading from
the critical path.  The widths are calculated as:

new_width = new_height * old_width / old_height

Where new_height = 48.  You can use "identify" to report an
image size (it's displayed as width x height).

Note: This commit does NOT resize the ACTUAL images served,
it *only* sets the width and height as will be displayed by
the browser.  We'll eventually resize the images to reduce the
number of bytes transmitted, but that is a separate task.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>