Fix mailto: link - spaces MUST be encoded as "%20", not "+"
Fix an oddity of "mailto" links; since they do not go to websites,
embedded spaces MUST be encoded as "%20", not as "+".
So replace all "+" characters as "%20"; this is a firm fix,
because "to_param" encodes "+" as "%2B" (so the *only* meaning
of "+" at this point is to encode a space).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>